Updated `load_scales_compat` to patch the model's `load_state_dict` to ignore missing scale factors
